requirements_v2.txt
# Minimal, pinned-major dependency set for the V2 pipeline.
# Install with:  pip install -r requirements_v2.txt

flask>=2.3
werkzeug>=2.3

numpy>=1.24
pandas>=2.0
pillow>=10.0
scikit-learn>=1.3
matplotlib>=3.7

# Face geometry
mediapipe>=0.10
opencv-python>=4.8

# Visual similarity (CLIP)
torch>=2.1
open-clip-torch>=2.24

# Legacy v1 path (optional — comment out if you only run V2)
face-recognition>=1.3
dlib>=19.24

# Dev / test
pytest>=7.4
